+++ BurnOutSharp/ProtectionType/Macrovision.CDilla.cs
@@ -10,10 +10,15 @@ using BurnOutSharp.Tools;
 namespace BurnOutSharp.ProtectionType
 {
     /// <summary>
-    /// This is a placeholder C-Dilla (sub-Macrovision) specific functionality
+    /// C-Dilla
+    /// https://archive.org/details/3ds-max-4.2original
+    /// https://knowledge.autodesk.com/support/3ds-max/troubleshooting/caas/sfdcarticles/sfdcarticles/Description-of-the-C-Dilla-License-Management-System.html
+    /// https://download.autodesk.com/mne/web/support/3dstudio/C-Dilla3.27.zip
+    /// https://forums.anandtech.com/threads/found-c-dilla-on-my-computer-get-it-off-get-it-off-heeeelp.857554/
+    /// https://www.extremetech.com/extreme/53108-macrovision-offers-closer-look-at-safecastcdilla
     /// </summary>
     public partial class Macrovision
     {
-        // TODO: Add methods for specific PE, path, and version checks here
+        // TODO: Add C-Dilla checks.
     }
 }
