+++ BurnOutSharp/ProtectionType/Macrovision.SafeCast.cs
@@ -10,7 +10,22 @@ using BurnOutSharp.Tools;
 namespace BurnOutSharp.ProtectionType
 {
     /// <summary>
-    /// This is a placeholder SafeCast (sub-Macrovision) specific functionality
+    /// SafeCast is in the same family of protections, and appears to mainly be for license management, and doesn't appear to affect the mastering of the disc in any way.
+    /// Although SafeCast is most commonly used in non-game software, there is one game that comes with both SafeDisc and SafeCast protections (Redump entry 83145).
+    /// Macrovision bought the company C-Dilla and created SafeCast based on C-Dilla's existing products (https://web.archive.org/web/20030212040047/http://www.auditmypc.com/freescan/readingroom/cdilla.asp).
+    /// That being said, there are references to C-Dilla within SafeDisc protected executables as early as 1.00.025, making the exact relationship between SafeDisc/Macrovision/C-Dilla unclear.
+    /// SafeCast UNSORTED samples:
+    /// https://www.virustotal.com/gui/file/220b8eec64d43fd854a48f28efa915c158a06f3675cbf0ef52796750984d9e1c/details
+    /// https://groups.google.com/g/alt.english.usage/c/kcBzeqXgE-M
+    /// https://www.shouldiremoveit.com/aerosofts-FDC-Live-Cockpit-64651-program.aspx
+    /// https://archive.org/details/danceejay4
+    /// https://archive.org/details/ejay_nestle_trial
+    /// https://archive.org/details/eJayXtremeSoundtraxx
+    /// SafeCast resources: 
+    /// https://web.archive.org/web/20031204024544mp_/http://www.macrovision.com/products/safecast/index.shtml
+    /// https://web.archive.org/web/20010417222834/http://www.macrovision.com/press_rel3_17_99.html
+    /// https://www.extremetech.com/computing/53394-turbotax-so-what-do-i-do-now/4
+    /// https://web.archive.org/web/20031013085038/http://www.pestpatrol.com/PestInfo/c/c-dilla.asp
     /// </summary>
     public partial class Macrovision
     {
@@ -24,6 +39,8 @@ namespace BurnOutSharp.ProtectionType
             string name = pex.FileDescription;
             if (name?.Equals("SafeCast2", StringComparison.OrdinalIgnoreCase) == true)
                 return $"SafeCast";
+
+            // Check for CDSHARE/DISAG_SH sections
 
             return null;
         }
@@ -55,6 +72,9 @@ namespace BurnOutSharp.ProtectionType
 
                 // Found in Redump entry 83145.
                 new PathMatchSet(new PathMatch("CDAC21BA.DLL", useEndsWith: true), "SafeCast"),
+
+                // Shown in multiple sources (such as https://groups.google.com/g/alt.english.usage/c/kcBzeqXgE-M) to be associated with SafeCast, but no samples have been found as of yet.
+                new PathMatchSet(new PathMatch("SCRfrsh.exe", useEndsWith: true), "SafeCast (Unconfirmed - Please report to us on Github)"),
             };
 
             return MatchUtil.GetFirstMatch(path, matchers, any: true);
